refactor(core): extract pure helpers from DesktopAssistantService into a focused module
About 120 lines of small, self-contained helpers were sitting at the
bottom of DesktopAssistantService:

- composeOneShotMessage, composePromoteMessage: prompt composition
- splitLeadingEmoji: emoji extraction for icon derivation
- clampLimit: pagination bound for /history
- readDesktopAssistantMeta + StoredDesktopAssistantMeta:
  workflow_entity.meta.desktopAssistant accessor
- extractBuiltWorkflowId, extractWorkflowLoopBuildOutcome: the two
  build-outcome detection paths (planned-task tasks-update heuristic
  and the workflow-loop thread-metadata reader)

None of these depend on DI or the database; they're deterministic
functions over request payloads, workflow JSON, and stored event-bus
state. Move them to desktop-assistant.helpers.ts and import where used.

splitLeadingEmoji's existing test now imports from the helpers file.
Behavior preserved; 60 tests still pass.

Service file goes from 718 \u2192 509 lines.
